rename: reader, writer를 모아둔 패키지명을 domain에서 adaptor로 변경
