+++ packy-api/src/main/java/com/dilly/admin/application/AdminService.java
@@ -3,12 +3,12 @@ package com.dilly.admin.application;
 import com.dilly.admin.dto.response.BoxImgResponse;
 import com.dilly.admin.dto.response.ImgResponse;
 import com.dilly.admin.dto.response.MusicResponse;
-import com.dilly.gift.domain.BoxReader;
-import com.dilly.gift.domain.EnvelopeReader;
-import com.dilly.gift.domain.MusicReader;
-import com.dilly.gift.domain.StickerReader;
+import com.dilly.gift.adaptor.BoxReader;
+import com.dilly.gift.adaptor.EnvelopeReader;
+import com.dilly.gift.adaptor.MusicReader;
+import com.dilly.gift.adaptor.StickerReader;
 import com.dilly.gift.dto.response.EnvelopeListResponse;
-import com.dilly.member.domain.ProfileImageReader;
+import com.dilly.member.adaptor.ProfileImageReader;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
