+++ packy-api/src/main/java/com/dilly/auth/application/AuthService.java
@@ -6,10 +6,10 @@ import static com.dilly.member.Provider.TEST;
 
 import com.dilly.auth.AppleAccount;
 import com.dilly.auth.KakaoAccount;
-import com.dilly.auth.domain.AppleAccountReader;
-import com.dilly.auth.domain.AppleAccountWriter;
-import com.dilly.auth.domain.KakaoAccountReader;
-import com.dilly.auth.domain.KakaoAccountWriter;
+import com.dilly.auth.adaptor.AppleAccountReader;
+import com.dilly.auth.adaptor.AppleAccountWriter;
+import com.dilly.auth.adaptor.KakaoAccountReader;
+import com.dilly.auth.adaptor.KakaoAccountWriter;
 import com.dilly.auth.dto.request.SignupRequest;
 import com.dilly.auth.dto.response.SignInResponse;
 import com.dilly.auth.model.AppleAccountInfo;
@@ -20,16 +20,16 @@ import com.dilly.exception.UnsupportedException;
 import com.dilly.global.utils.SecurityUtil;
 import com.dilly.jwt.JwtService;
 import com.dilly.jwt.RefreshToken;
-import com.dilly.jwt.domain.JwtReader;
-import com.dilly.jwt.domain.JwtWriter;
+import com.dilly.jwt.adaptor.JwtReader;
+import com.dilly.jwt.adaptor.JwtWriter;
 import com.dilly.jwt.dto.JwtResponse;
 import com.dilly.member.Member;
 import com.dilly.member.ProfileImage;
 import com.dilly.member.Provider;
 import com.dilly.member.Status;
-import com.dilly.member.domain.MemberReader;
-import com.dilly.member.domain.MemberWriter;
-import com.dilly.member.domain.ProfileImageReader;
+import com.dilly.member.adaptor.MemberReader;
+import com.dilly.member.adaptor.MemberWriter;
+import com.dilly.member.adaptor.ProfileImageReader;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
