+++ packy-api/src/main/java/com/dilly/gift/adaptor/ReceiverReader.java
@@ -1,22 +1,19 @@
-package com.dilly.gift.domain;
+package com.dilly.gift.adaptor;
 
 import com.dilly.gift.GiftBox;
 import com.dilly.gift.Receiver;
 import com.dilly.gift.dao.ReceiverRepository;
-import com.dilly.member.Member;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class ReceiverWriter {
+public class ReceiverReader {
 
     private final ReceiverRepository receiverRepository;
 
-    public void save(Member member, GiftBox giftBox) {
-        receiverRepository.save(Receiver.builder()
-            .receiver(member)
-            .giftBox(giftBox)
-            .build());
+    public List<Receiver> findByGiftBox(GiftBox giftBox) {
+        return receiverRepository.findByGiftBox(giftBox);
     }
 }
