+++ packy-api/src/main/java/com/dilly/gift/adaptor/ReceiverWriter.java
@@ -1,19 +1,22 @@
-package com.dilly.gift.domain;
+package com.dilly.gift.adaptor;
 
 import com.dilly.gift.GiftBox;
 import com.dilly.gift.Receiver;
 import com.dilly.gift.dao.ReceiverRepository;
-import java.util.List;
+import com.dilly.member.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class ReceiverReader {
+public class ReceiverWriter {
 
     private final ReceiverRepository receiverRepository;
 
-    public List<Receiver> findByGiftBox(GiftBox giftBox) {
-        return receiverRepository.findByGiftBox(giftBox);
+    public void save(Member member, GiftBox giftBox) {
+        receiverRepository.save(Receiver.builder()
+            .receiver(member)
+            .giftBox(giftBox)
+            .build());
     }
 }
