+++ packy-api/src/main/java/com/dilly/gift/application/GiftService.java
@@ -10,17 +10,17 @@ import com.dilly.gift.GiftBox;
 import com.dilly.gift.GiftType;
 import com.dilly.gift.Letter;
 import com.dilly.gift.Receiver;
-import com.dilly.gift.domain.BoxReader;
-import com.dilly.gift.domain.EnvelopeReader;
-import com.dilly.gift.domain.GiftBoxReader;
-import com.dilly.gift.domain.GiftBoxStickerReader;
-import com.dilly.gift.domain.GiftBoxStickerWriter;
-import com.dilly.gift.domain.GiftBoxWriter;
-import com.dilly.gift.domain.LetterWriter;
-import com.dilly.gift.domain.PhotoReader;
-import com.dilly.gift.domain.PhotoWriter;
-import com.dilly.gift.domain.ReceiverReader;
-import com.dilly.gift.domain.ReceiverWriter;
+import com.dilly.gift.adaptor.BoxReader;
+import com.dilly.gift.adaptor.EnvelopeReader;
+import com.dilly.gift.adaptor.GiftBoxReader;
+import com.dilly.gift.adaptor.GiftBoxStickerReader;
+import com.dilly.gift.adaptor.GiftBoxStickerWriter;
+import com.dilly.gift.adaptor.GiftBoxWriter;
+import com.dilly.gift.adaptor.LetterWriter;
+import com.dilly.gift.adaptor.PhotoReader;
+import com.dilly.gift.adaptor.PhotoWriter;
+import com.dilly.gift.adaptor.ReceiverReader;
+import com.dilly.gift.adaptor.ReceiverWriter;
 import com.dilly.gift.dto.request.GiftBoxRequest;
 import com.dilly.gift.dto.response.BoxResponse;
 import com.dilly.gift.dto.response.EnvelopeResponse;
@@ -31,7 +31,7 @@ import com.dilly.gift.dto.response.PhotoResponse;
 import com.dilly.gift.dto.response.StickerResponse;
 import com.dilly.global.utils.SecurityUtil;
 import com.dilly.member.Member;
-import com.dilly.member.domain.MemberReader;
+import com.dilly.member.adaptor.MemberReader;
 import java.util.Comparator;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
