+++ packy-api/src/test/java/com/dilly/global/IntegrationTestSupport.java
@@ -11,6 +11,10 @@ import com.dilly.gift.Letter;
 import com.dilly.gift.Photo;
 import com.dilly.gift.Receiver;
 import com.dilly.gift.Sticker;
+import com.dilly.gift.adaptor.GiftBoxWriter;
+import com.dilly.gift.adaptor.PhotoWriter;
+import com.dilly.gift.adaptor.ReceiverReader;
+import com.dilly.gift.adaptor.ReceiverWriter;
 import com.dilly.gift.application.GiftService;
 import com.dilly.gift.dao.BoxRepository;
 import com.dilly.gift.dao.EnvelopeRepository;
@@ -22,10 +26,6 @@ import com.dilly.gift.dao.PhotoRepository;
 import com.dilly.gift.dao.ProfileImageRepository;
 import com.dilly.gift.dao.ReceiverRepository;
 import com.dilly.gift.dao.StickerRepository;
-import com.dilly.gift.domain.GiftBoxWriter;
-import com.dilly.gift.domain.PhotoWriter;
-import com.dilly.gift.domain.ReceiverReader;
-import com.dilly.gift.domain.ReceiverWriter;
 import com.dilly.member.Member;
 import com.dilly.member.MemberRepository;
 import com.dilly.mypage.application.MyPageService;
